server_test/requirements.txt
httpx[http2]>=0.27
pytest>=8.2
pytest-asyncio>=0.23
anyio>=4.3
python-dotenv>=1.0
asyncpg>=0.29
bcrypt>=4.2
